course-app-easy-2 backend done

+++ week-3/assignment/course-app-easy/index.js
@@ -5,28 +5,31 @@ const fs = require('fs')
 
 app.use(express.json());
 
-function authenticateUser(username, password, adminList) {
-  for (const user of adminList) {
-    if (user.username === username && user.password === password) {
-      return true; // User is authenticated
+function authenticateUser(req,res,next){
+  const {username, password} = req.headers
+  fs.readFile('userProfile.json', 'utf-8', (err, data) => {
+    if (err) throw err;
+    const userList = JSON.parse(data);
+    const user = userList.find(a => a.username == username && a.password == password)
+    if(user){
+      req.user = user //updating req attribute 
+      next()
+    } else{
+      res.status(403).json({'message' : 'User authentication failed'})
     }
-  }
-  return false; // User is not authenticated
+  })
 }
 
 function authenticateAdmin(req,res,next){
-  var username = req.headers.username
-  var password = req.headers.password
+  const {username, password} = req.headers
   fs.readFile('adminProfile.json', 'utf-8', (err, data) => {
     if (err) throw err;
-
     const adminList = JSON.parse(data);
-    const authenticated = authenticateUser(username, password, adminList);
-
-    if(authenticated){
+    const admin = adminList.find(a => a.username == username && a.password == password)
+    if(admin){
       next()
     } else{
-      res.status(404).send("invalid user")
+      res.status(403).json({'message' : 'Admin authentication failed'})
     }
   })
 }
@@ -41,11 +44,16 @@ app.post('/admin/signup', (req, res) => {
   fs.readFile('adminProfile.json','utf-8', (err,data)=>{
     if (err) throw err
     const adminProfile = JSON.parse(data)
-    adminProfile.push(newAdmin)
-    fs.writeFile('adminProfile.json',JSON.stringify(adminProfile),(err)=>{
-      if (err) throw err
-      res.status(201).send("new admin added succesfully")
+    const existingAdmin = adminProfile.find(a => a.username == req.body.username)
+    if(existingAdmin){
+      res.status(403).send("admin already exists")
+    } else{
+      adminProfile.push(newAdmin)
+      fs.writeFile('adminProfile.json',JSON.stringify(adminProfile),(err)=>{
+        if (err) throw err
+        res.status(201).send("new admin added succesfully")
     })
+    }
   })
 });
 
@@ -74,34 +82,27 @@ app.post('/admin/courses', authenticateAdmin, (req, res) => {
 });
 
 app.put('/admin/courses/:courseId', authenticateAdmin, (req, res) => {
-  const userID = req.params.courseId
-  console.log(userID)
-  fs.readFile('courses.json', 'utf-8', (err,data)=>{
-    if (err) throw err
-    var courseList = JSON.parse(data)
-    var counter = 0
-    for (course of courseList){
-      if(userID == course.courseID){
-        const updatedCourse = {
-          title : req.body.title,
-          description : req.body.description,
-          price : req.body.price,
-          imageLink : req.body.imageLink,
-          published: false,
-          courseID : userID
+  const courseID = parseInt(req.params.courseId);
+  fs.readFile('courses.json', 'utf-8', (err, data) => {
+    if (err) throw err;
+    var courseList = JSON.parse(data);
+    const course = courseList.find(a => a.courseID === courseID);
+    if (course) {
+      Object.assign(course, req.body); //Object.assign copies values from req.body and updates them in course
+      fs.writeFile('courses.json', JSON.stringify(courseList), (err) => {
+        if (err) {
+          console.error(err);
+          res.status(500).json({ message: 'Failed to update course' });
+        } else {
+          res.json({ message: 'Course updated successfully' });
         }
-        courseList[counter] = updatedCourse
-        break
-      }
-      counter ++
+      });
+    } else {
+      res.status(404).json({ message: 'Course not found' });
     }
-    fs.writeFile('courses.json', JSON.stringify(courseList),(err)=>{
-      if (err) throw err
-      res.status(201).send("updated successfully")
-    })
-    
-  })
+  });
 });
+
 
 app.get('/admin/courses', authenticateAdmin, (req, res) => {
   fs.readFile('courses.json', 'utf-8', (err,data)=>{
@@ -113,12 +114,8 @@ app.get('/admin/courses', authenticateAdmin, (req, res) => {
 
 // User routes
 app.post('/users/signup', (req, res) => {
-  const newUser = {
-    id : Math.floor(Math.random()*1000000),
-    username : req.body.username,
-    password : req.body.password,
-    courses : []
-  }
+  const newUser ={...req.body, id: Math.floor(Math.random()*1000000), courses : []}
+  //The spread syntax ... is used to copy all enumerable properties from the req.body object into the new user object
   fs.readFile('userProfile.json','utf-8', (err,data)=>{
     if (err) throw err
     const userProfile = JSON.parse(data)
@@ -130,54 +127,36 @@ app.post('/users/signup', (req, res) => {
   })
 });
 
-app.post('/users/login', (req, res) => {
-  const username = req.headers.username;
-  const password = req.headers.password;
-
-  fs.readFile('userProfile.json', 'utf-8', (err, data) => {
-    if (err) throw err;
-
-    const adminList = JSON.parse(data);
-    const authenticated = authenticateUser(username, password, adminList);
-
-    if (authenticated) {
-      res.status(200).send("Logged in successfully");
-    } else {
-      res.status(401).send("User not found or invalid credentials");
-    }
-  })
+app.post('/users/login', authenticateUser, (req, res) => {
+  res.send("logged in succesfully")
 });
 
-app.get('/users/courses', (req, res) => {
+app.get('/users/courses', authenticateUser, (req, res) => {
   fs.readFile('courses.json', 'utf-8', (err,data)=>{
     if (err) throw err
     const courseList = JSON.parse(data)
-    res.json(courseList)
+    res.json({courses: courseList.filter(c => c.published)})
   })
 });
 
-app.post('/users/courses/:courseId', (req, res) => {
-  const courseID = req.params.courseId
-  const username = req.headers.username
-  var purchasedCourse
+app.post('/users/courses/:courseId',authenticateUser, (req, res) => {
+  const courseID = parseInt(req.params.courseId)
   fs.readFile('userProfile.json', 'utf-8', (err,data)=>{
     if (err) throw err
     var userList = JSON.parse(data)
     fs.readFile('courses.json','utf-8', (err,data)=>{
       if (err) throw err
       const courseList = JSON.parse(data)
-      for (course of courseList){
-        if(course.courseID == courseID){
-          purchasedCourse = course
+      const course = courseList.find(c => c.courseID == courseID && c.published)
+      if(course){
+        var user = userList.find(a => a.username == req.user.username)
+        if(user){
+          user.courses.push(course.courseID)
+        } else {
+          res.status(404).send("user not found")
         }
-      }
-      var counter = 0
-      for (user of userList){
-        if(username == user.username){
-          userList[counter].courses.push(purchasedCourse)
-          break
-        }
-      counter ++
+      } else{
+        res.status(404).send("course not found")
       }
       fs.writeFile('userProfile.json',JSON.stringify(userList), (err)=>{
         if (err) throw err
@@ -187,19 +166,18 @@ app.post('/users/courses/:courseId', (req, res) => {
   })
 });
 
-app.get('/users/purchasedCourses', (req, res) => {
+app.get('/users/purchasedCourses', authenticateUser, (req, res) => {
   var username = req.headers.username
   fs.readFile('userProfile.json', 'utf-8', (err,data)=>{
     if (err) throw err
-    const userProfile = JSON.parse(data)
-    let counter = 0
-    for(user of userProfile){
-      if (username == user.username){
-        var purchasedCourses = userProfile[counter].courses
-        res.json(purchasedCourses)
-      }
-      counter ++
-    } 
+    const userList = JSON.parse(data)
+    fs.readFile('courses.json','utf-8',(err,data)=>{
+      if (err) throw err
+      const courseList = JSON.parse(data)
+      var purchasedCourses = courseList.filter(c => req.user.courses.includes(c.courseID))
+      console.log(purchasedCourses)
+      res.json(purchasedCourses)
+    })
   })
 });
